requirements.txt
Flask
Flask-PyMongo
PyJWT
Werkzeug
python-dotenv
Flask-CORS
gunicorn